GBE-679: Code cleanup: broke out a number of classes from gbe/models/remains.py
