+++ expo/gbe/models/persona.py
@@ -1,0 +1,43 @@
+from django.db.models import ForeignKey
+from profile import Profile
+from performer import Performer
+
+class Persona (Performer):
+    '''
+    The public persona of one person who performs or teaches.
+    May be aggregated into a group or a troupe,
+    or perform solo, or both. A single person might conceivably maintain two
+    distinct performance identities and therefore have multiple
+    Persona objects associated with their profile. For example, a
+    person who dances under one name and teaches under another would
+    have multiple Personae.
+    performer_profile is the profile of the user who dons this persona.
+    '''
+    performer_profile = ForeignKey(Profile, related_name="personae")
+
+    '''
+    Returns the single profile associated with this persona
+    '''
+    def get_profiles(self):
+        return [self.performer_profile]
+
+    def get_schedule(self):
+        return self.performer_profile.get_schedule()
+
+    @property
+    def complete(self):
+        return (self.performer_profile.complete and
+                self.name is not '' and
+                self.bio is not '')
+
+    # def append_alerts(self, alerts):
+    #     '''
+    #     Find any alerts generated by this object's data and append them to
+    #     the alerts dict presented as a parameter
+    #     '''
+    #     alerts = super(Persona, self).append_alerts(alerts)
+    #     return alerts
+
+    class Meta:
+        verbose_name_plural = 'personae'
+        app_label = "gbe"
