+++ expo/scheduler/models.py
@@ -548,6 +548,8 @@ class EventItem (models.Model):
         self.visible = False
         self.save()
 
+
+
     @property
     def payload(self):
         return self.sched_payload
